Creation of update form
Original intent: (incorrect) using a for loop and innerhtml to add
elements to html file. Does NOT work in Meteor!

Amendment (currently working on): using a template for the form, and a
template for the Reasons part if the printer is out of order. (Inspired
by Eni’s meetingApp)

Methods changed/created: click .selectedStatus, click .otherSelected

Helpers: openForm, printerIsOutOfOrder, Session.setDefault (global var)